refactor(tests): move `subscript_of()` into the module the sweeps share
