+++ tests/unit/name_resolution.py
@@ -16,13 +16,15 @@
 #  You should have received a copy of the GNU Lesser General Public License
 #  along with Pyrogram.  If not, see <http://www.gnu.org/licenses/>.
 
-"""Reading a dotted name against the package, shared by the docstring and annotation sweeps.
+"""What the sweeps over the package share: which files they read, and how they read a name.
 
-Both ask the same question of a different kind of text: does this name exist. Neither can
-answer it statically, because the tree is one import cycle wide and half the names it
-writes are only bound under `TYPE_CHECKING`, so both resolve against the imported package.
+The name sweeps ask the same question of a different kind of text: does this name exist.
+Neither can answer it statically, because the tree is one import cycle wide and half the
+names it writes are only bound under `TYPE_CHECKING`, so both resolve against the imported
+package.
 """
 
+import ast
 import importlib
 import pathlib
 from typing import Any, Final, Iterator, Sequence
@@ -33,6 +35,21 @@ PACKAGE_ROOT: Final[pathlib.Path] = REPOSITORY_ROOT / "pyrogram"
 # The generated tree is rewritten wholesale by `make api` from the TL schema, so a repair
 #  there lives until the next schema update and no longer.
 GENERATED_TREE: Final[pathlib.Path] = PACKAGE_ROOT / "raw"
+
+
+def subscript_of(node: ast.Subscript) -> ast.expr:
+    """What sits inside `X[...]`, whichever way this Python spells it.
+
+    Python 3.9 stopped wrapping a subscript in `ast.Index`, and 3.8 is still supported.
+    The wrapper is read by name rather than by `isinstance`, because `ast.Index` has been
+    deprecated since 3.9 and touching it raises a `DeprecationWarning`.
+    https://docs.python.org/3/whatsnew/3.9.html#deprecated
+    """
+    inner = node.slice
+
+    # TODO: Delete the unwrapping, and the test that covers it, when the floor moves off
+    #       3.8. Nothing else reads `ast.Index`, and 3.9 stopped emitting it.
+    return getattr(inner, "value", inner) if inner.__class__.__name__ == "Index" else inner
 
 
 def hand_written_files() -> Iterator[pathlib.Path]:
